requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
python-jose==3.5.0
email-validator==2.2.0
